staging: unisys: removed unused 'visor_device.respond_to_device_create'
(I cherry-picked this onto my dev branch because this is in the
pipeline for staging-next, but hasn't made it there yet.  This WILL
be in staging-next by the time my next commit is submitted.)

The respond_to_device_create flag was used previously when we used to
delay responses to create requests until the drivers were finished
loading. This behaviour was removed some time ago, yet the field still
existed, while never being referenced or even initialized.

Github issue #19

Signed-off-by: Alexander Curtin <alexander.curtin@unisys.com>