DO NOT SUBMIT YET: staging: unisys: added display string size define
Display string size that is guaranteed to be no longer the 99 characters.

Don't use a magic number throughout the code. It needs to be defined once
and used in the locations.

This patch solves an issue discovered by sonarqube:
Extract this magic number '99' into a constant, variable declaration
or an enum.

Signed-off-by: Erik Arfvidson <erik.arfvidson@unisys.com>
Signed-off-by: David Kershner <david.kershner@unisys.com>