staging: unisys: visornic: prevent double-unlock of priv_lock
Previously, devdata->priv_lock was being unlocked in visornic_serverdown()
both before calling visornic_serverdown_complete(), then again at the end
of the function.  This bug was corrected.

Simplified the structure of visornic_serverdown() to make it easier to
follow and decrease chances of bugs. The main-path logic now falls through
down the left-side of the page, with a common error-exit point to handle
error conditions.

Signed-off-by: Tim Sell <Timothy.Sell@unisys.com>
Signed-off-by: David Kershner <david.kershner@unisys.com>