staging: unisys: visornic: prevent double-unlock of priv_lock
Previously, devdata->priv_lock was being unlocked in visornic_serverdown()
both before calling visornic_serverdown(), then again at the end of the
function.  This bug was corrected.

The structure of visornic_serverdown() was also improved so as to make it
easier to follow, to decrease the chance that such bugs will be introduced
again.  The main-path logic now falls thru down the left-side of the page,
with a common error-exit point to handle error conditions.

Signed-off-by: Tim Sell <Timothy.Sell@unisys.com>