staging: unisys: visorbus: fix race condition involving interrupt enable
The observed race condition involved I/Os issued by visorhba_probe(),
(which is called via visorbus visordriver_probe_device(), which is typically
called from the device_add() done in create_visor_device()), and the setting
of the channel state (visorbus_set_channel_state()) to CHANNELCLI_ATTACHED.

The IO partition will not provide interrupts to a device whose channel state
is not CHANNELCLI_ATTACHED or CHANNELCLI_OWNED.  So in cases where a visor
driver's probe() function expected interrupts AND required completed I/Os from
the device, we could get into a hang scenario where we would be waiting
forever for I/Os to complete.

The problem was fixed by ensuring that the channel state is set to
CHANNELCLI_ATTACHED *before* visordriver_probe_device() is called.  This
ensures that the device's probe() function will be able to complete I/Os
with interrupts if it needs to.

Without this patch, in a single-proc PC test environment, I could only boot
successfully about 1 in every 3 times with a visorhba boot device.  On the
occasions when it would not boot up, we would fail to boot because we couldn't
access the boot device, and the IO partition would show this for the visorhba
boot device:

    CliStateOS: 0

A "CliStateOS" 0 is CHANNELCLI_DETACHED, and when in this state the IO
will not inject interrupts for the device.  In the working cases, the IO
partition would show this for the visorhba boot device:

    CliStateOS: 3

A "CliStateOS" 3 is CHANNELCLI_ATTACHED, which is a normal operational state
for the device.

I booted 10 times-in-a-row successfully with a visorhba boot device to
validate this patch.

Signed-off-by: Tim Sell <Timothy.Sell@unisys.com>