staging: unisys: visorinput: fix bug revealed by commit 1af5e9719bc18a9a7e68563e992d57af8ebec2c0
...and referenced within github issue #58.

I have regression-tested this on a vanilla for-later branch, but haven't yet tried
it in conjunction with commit 1af5e9719bc18a9a7e68563e992d57af8ebec2c0 and
friends.

The bug resulted in calls visorbus_disable_channel_interrupts() not doing the
right thing, which for the polling (irq == 0) case, would result in a subsequent
call to visorbus_enable_channel_interrupts() to fail.