staging: unisys: visorbus: relocate error-check from isr to registration
It just makes more sense to do the NULL-pointer check when interrupts are
enabled, rather than on *every* interrupt.

This is the staging-next based version of the fix for github #65.
It compiles cleanly.

But our current plan is to ONLY release this for fix in for-later.  I just
wanted to stash this patch in case we need it for some reason.

Signed-off-by: Tim Sell <Timothy.Sell@unisys.com>