staging: unisys: visorbus: change how we know when we are in irq mode
Previously, the code was using "struct visor_device.irq != 0" to determine
whether or not we were in irq mode, and forcing visor_device.irq to 0 in
visorbus_register_for_channel_interrupts() if it failed.  But because the irq
field is not otherwise set in visorbus_register_for_channel_interrupts() (it
is in fact assigned earlier when the device is initially discovered, in
my_device_create() of visorchipset.c), this was somewhat non-orthogonal.

Also, this prior behavior would have also prevented the possibility of forcing
subsequent attempts at request_irq() by unloading+reloading the function driver
controlling the device.

We now use struct visor_device.interrupt_recv_queue to indicate whether or not
we are operating in irq mode, since it is the job of
visorbus_register_for_channel_interrupts() to set that field.

* interrupt_recv_queue == -1 indicates NOT in irq mode
* interrupt_recv_queue >=  0 indicates in irq mode
  (using the value as the channel receive queue)

The VISORBUS_IRQ_INTERRUPTS_ENABLED macro just makes this more obvious.

visorbus_register_for_channel_interrupts() also now properly cleans up by
calling free_irq() when necessary.

Signed-off-by: Tim Sell <Timothy.Sell@unisys.com>