staging: unisys: visorbus: Remove notifier-related code from visorbus
When this functionality was first implemented, visorchipset and visorbus
were separate drivers, which necessitated a registration mechanism for
them to communicate.  More-recently, visorchipset and visorbus were
combined into a single driver, and now exist as separate source files
within the same driver, known as 'visorbus'.  This eliminated the need
for a registration mechanism, but it has remained nevertheless until now.
For the sake of simplification, this registration mechanism is now being
removed. This patch removes the "notifier" portion of the
"notifier/responder" paradigm that existed back when visorbus and
visorchipset were two distinct drivers.

Signed-off-by: David Binder <david.binder@unisys.com>